登陆处理

+++ trunk/crm/application/libs/Auth.php
@@ -15,8 +15,18 @@ class Auth
         // if user is still not logged in, then destroy session, handle user as "not logged in" and
         // redirect user to login page
         if (!isset($_SESSION['user_logged_in'])) {
-            Session::destroy();
-            header('location: ' . URL . 'login/unloginresponse');
+        	// user has remember-me-cookie ? then try to login with cookie ("remember me" feature)
+        	$login_successful = false;
+        	if (isset($_COOKIE['rememberme'])) {
+        		$controller = new Controller();
+        		$login_model = $controller->loadModel('Login');
+        		$login_successful = $login_model->loginWithCookie();
+        	}
+        	if(!$login_successful)
+        	{
+            	Session::destroy();
+            	header('location: ' . URL . 'login/unloginresponse');
+        	}
         }
     }
 }
